Wiki_ChatBot/requirements.txt
langchain==0.0.154
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.25.0
faiss-cpu==1.7.4
altair<5